requirements.txt
flask
Flask-SQLAlchemy
flask-login
markdown
python-dotenv